+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 2.8.0
 - New control parameter 'subdailyCarbonBalance', applying to growth simulations with transpiration = "Sperry"
 - Hydraulic redistribution implemented for transpiration = "Granier"
+- Growth with "Sperry" using canopy temperature
 
 # Version 2.7.7
 - Imputation relationship between RGRcambiummax and SRsapwood